@@ -1,6 +1,6 @@
 package com.techyourchance.unittesting.screens.common.controllers;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.techyourchance.unittesting.common.CustomApplication;
 import com.techyourchance.unittesting.common.dependencyinjection.ControllerCompositionRoot;
