@@ -1,7 +1,8 @@
 package com.techyourchance.unittesting.screens.common.fragmentframehelper;
 
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.Nullable;
+
+import androidx.fragment.app.Fragment;
 
 public interface HierarchicalFragment {
     /**
