@@ -1,7 +1,7 @@
 package com.techyourchance.unittesting.screens.common.main;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.widget.FrameLayout;
 
 import com.techyourchance.unittesting.screens.common.controllers.BackPressDispatcher;
