@@ -1,15 +1,14 @@
 package com.techyourchance.unittesting.screens.common.navdrawer;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.NavigationView;
-import android.support.v4.widget.DrawerLayout;
+import androidx.annotation.Nullable;
 import android.view.Gravity;
 import android.view.LayoutInflater;
-import android.view.MenuItem;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
 
+import androidx.drawerlayout.widget.DrawerLayout;
+
+import com.google.android.material.navigation.NavigationView;
 import com.techyourchance.unittesting.R;
 import com.techyourchance.unittesting.screens.common.views.BaseObservableViewMvc;
 
@@ -18,32 +17,26 @@ public class NavDrawerViewMvcImpl extends BaseObservableViewMvc<NavDrawerViewMvc
 
     private final DrawerLayout mDrawerLayout;
     private final FrameLayout mFrameLayout;
-    private final NavigationView mNavigationView;
 
     public NavDrawerViewMvcImpl(LayoutInflater inflater, @Nullable ViewGroup parent) {
         setRootView(inflater.inflate(R.layout.layout_drawer, parent, false));
         mDrawerLayout = findViewById(R.id.drawer_layout);
         mFrameLayout = findViewById(R.id.frame_content);
-        mNavigationView = findViewById(R.id.nav_view);
+        NavigationView mNavigationView = findViewById(R.id.nav_view);
 
-        mNavigationView.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
-            @Override
-            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                mDrawerLayout.closeDrawers();
-                if (item.getItemId() == R.id.drawer_menu_questions_list) {
-                    for (Listener listener : getListeners()) {
-                        listener.onQuestionsListClicked();
-                    }
+        mNavigationView.setNavigationItemSelectedListener(item -> {
+            mDrawerLayout.closeDrawers();
+            if (item.getItemId() == R.id.drawer_menu_questions_list) {
+                for (Listener listener : getListeners()) {
+                    listener.onQuestionsListClicked();
                 }
-                return false;
             }
+            return false;
         });
     }
 
     @Override
-    public void openDrawer() {
-        mDrawerLayout.openDrawer(Gravity.START);
-    }
+    public void openDrawer() { mDrawerLayout.openDrawer(Gravity.START); }
 
     @Override
     public boolean isDrawerOpen() {
