@@ -1,12 +1,13 @@
 package com.techyourchance.unittesting.screens.questionslist;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
 
 import com.techyourchance.unittesting.screens.common.controllers.BaseFragment;
 
