@@ -1,8 +1,8 @@
 package com.techyourchance.unittestinginandroid.example14;
 
 import android.app.Activity;
-import android.os.Bundle;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.Nullable;
 
 public class MyActivity extends Activity {
 
