@@ -1,11 +1,12 @@
 package com.techyourchance.unittesting.screens.questiondetails;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.techyourchance.unittesting.questions.FetchQuestionDetailsUseCase;
 import com.techyourchance.unittesting.questions.QuestionDetails;
